[SCB-1841]minor fixes: package name refactor and delete unused code
