+++ LICENSE
@@ -210,7 +210,7 @@ following licenses.
 
 
 ================================================================
-For foundations/foundation-vertx/src/main/java/io/vertx/ext/web/impl/MimeTypesUtils.java
+For foundations/foundation-common/src/main/java/org/apache/servicecomb/foundation/common/utils/MimeTypesUtils.java
     transports/transport-rest/transport-rest-vertx/src/main/java/org/apache/servicecomb/transport/rest/vertx/RestBodyHandler.java
 ================================================================
 This product bundles files from vertx which is licensed under the Apache License v2.
