+++ integration-tests/it-producer-deploy-springboot2-standalone/pom.xml
@@ -72,36 +72,6 @@
             <target>1.8</target>
           </configuration>
         </plugin>
-        <plugin>
-          <groupId>org.apache.rat</groupId>
-          <artifactId>apache-rat-plugin</artifactId>
-          <configuration>
-            <excludes>
-              <exclude>DISCLAIMER</exclude>
-              <exclude>.travis.yml</exclude>
-              <exclude>**/*.md</exclude>
-              <excldue>**/*.MD</excldue>
-              <exclude>**/target/*</exclude>
-              <!-- Skip the code style configuration file -->
-              <exclude>**/etc/eclipse-java-google-style.xml</exclude>
-              <exclude>**/etc/intellij-java-google-style.xml</exclude>
-              <!-- Skip the ssl configuration files -->
-              <exculde>**/resources/ssl/**</exculde>
-              <!-- SKip the licenses files -->
-              <exclude>**/src/release/licenses/**</exclude>
-              <!-- Skip the demo log files -->
-              <exclude>**/logs/**.log</exclude>
-              <exclude>**/**.log</exclude>
-              <!-- Skip the target files -->
-              <exclude>**/target/**</exclude>
-              <!-- Skip the source files which are forked from vertx -->
-              <exclude>**/io/vertx/ext/web/impl/MimeTypesUtils.java</exclude>
-              <exclude>**/java/org/apache/servicecomb/transport/rest/vertx/RestBodyHandler.java</exclude>
-              <!--Skip protobuf generated file-->
-              <exclude>**/java/org/apache/servicecomb/foundation/protobuf/internal/model/ProtobufRoot.java</exclude>
-            </excludes>
-          </configuration>
-        </plugin>
       </plugins>
     </pluginManagement>
   </build>
