+++ pom.xml
@@ -228,8 +228,7 @@
             <exclude>**/resources/webroot/images/*.png</exclude>
             <exclude>**/resources/webroot/images/*.jpg</exclude>
             <!-- Skip the source files which are forked from vertx -->
-            <exclude>**/io/vertx/ext/web/impl/MimeTypesUtils.java</exclude>
-            <exclude>**/io/vertx/core/impl/VertxImpl.java</exclude>
+            <exclude>**/org/apache/servicecomb/foundation/common/utils/MimeTypesUtils.java</exclude>
             <exclude>**/java/org/apache/servicecomb/transport/rest/vertx/RestBodyHandler.java</exclude>
             <!--Skip protobuf generated file-->
             <exclude>**/java/org/apache/servicecomb/foundation/protobuf/internal/model/ProtobufRoot.java</exclude>
